==> code/01_Cleaning_Initial.ipynb <==
#This notebook reads in our dataframes, and returns a trimmed dataframe. The goal is to have dataframes that do not have any Nans or '[removed]' posts, so that we
#can analyze them.

import pandas as pd
import numpy as np

#Bring in the two dataframes
lpt_raw = pd.read_csv('./data/lpt_df_2.csv', index_col=0)
ulpt_raw = pd.read_csv('./data/ulpt_df_2.csv', index_col = 0)

#Investigating my dataframes:
lpt_raw.info()

#Checking for nan values that may have slipped through:
lpt_raw[['selftext', 'title']].isna().value_counts()

ulpt_raw[['selftext', 'title']].isna().value_counts()

#Dropping and verifying the new length:
lpt_raw.dropna(subset = ['selftext', 'title'], inplace=True)
len(lpt_raw)

ulpt_raw.dropna(subset = ['selftext', 'title'], inplace=True)
len(ulpt_raw)

#Check the length of each datafram without posts that are removed:
len(lpt_raw[lpt_raw['selftext'] != '[removed]'])

len(ulpt_raw[ulpt_raw['selftext'] != '[removed]'])

#7500 posts each! Jackpot!

#Creating our new trimmed dataframes:
lpt_trimmed = lpt_raw[lpt_raw['selftext'] != '[removed]']
ulpt_trimmed = ulpt_raw[ulpt_raw['selftext'] != '[removed]']

#Triple checking our length:
len(lpt_trimmed)

len(ulpt_trimmed)

#Exporting to the data folder:
lpt_trimmed.to_csv('./data/lpt_trimmed.csv')
ulpt_trimmed.to_csv('./data/ulpt_trimmed.csv')

#End Result is that I have two dataframes _trimmed to ~ 7500 posts each. Hooray!
