==> code/02_Cleaning_Final.ipynb <==
#In this notebook I am going to start with the dataframes that have been trimmed,
#The nans and the posts marked '[removed]' have been removed so we start with better data.

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt

lpt_raw = pd.read_csv('./data/lpt_trimmed.csv')
ulpt_raw = pd.read_csv('./data/ulpt_trimmed.csv')

#Probably not necessary, but just in case lets check:
lpt_raw.info()

ulpt_raw.info()

#Both dataframes have full title, selftext, and subreddit columns. This is where I will start.
lpt_df = lpt_raw[['title', 'selftext', 'subreddit']]
ulpt_df = ulpt_raw[['title', 'selftext', 'subreddit']]
lpt_df.head()

ulpt_df.head()

It has come to my attention that each of the posts in the subreddits start with either LPT or ULPT. This is an issue, as it will essentially flag the post right away when my classifier wants to check if it is LPT or ULPT. I need to deal with this somehow:

len(lpt_df[lpt_df['title'].str.contains('LPT')])/len(lpt_df)

len(ulpt_df[(ulpt_df['title'].str.contains('LPT' or 'ULPT'))])/len(ulpt_df)

#99% of LPTs and 96% of ULPTs have the respective word in the title. Let's remove them.

#looks like we might want to strip lpt and ulpt from the dataframes - here we use regex to get all instances of it in the title and post text:
lpt_df['title'] = lpt_df['title'].str.lower().str.strip(r'?u[lpt]')
lpt_df['selftext'] = lpt_df['selftext'].str.lower().str.strip(r'?u[lpt]')

Somehow my regex only caught most of them:

len(lpt_df[lpt_df['selftext'].str.contains('lpt')])

So let's strip again:

ulpt_df['title'] = ulpt_df['title'].str.lower().str.strip(r'?u[lpt]')
ulpt_df['selftext'] = ulpt_df['selftext'].str.lower().str.strip(r'?u[lpt]')

ULPT threads also contain some lpt references. We might want to remove these as they probably will confuse my future algorithm, and in the real world attempting to identify unethical text, there is a low probability that it will be coming from the LPT or ULPT community on Reddit anyway:

len(ulpt_df[ulpt_df['selftext'].str.contains('lpt')])

#Let's drop all the instances of lpt and ulpt we can find:
lpt_df.drop(lpt_df[lpt_df['title'].str.contains(r'(?i)lpt')].index, inplace=True)
lpt_df.drop(lpt_df[lpt_df['selftext'].str.contains(r'(?i)lpt')].index, inplace=True)
ulpt_df.drop(ulpt_df[ulpt_df['title'].str.lower().str.contains('ulpt')].index, inplace=True)
ulpt_df.drop(ulpt_df[ulpt_df['selftext'].str.lower().str.contains('ulpt')].index, inplace=True)
ulpt_df.drop(ulpt_df[ulpt_df['title'].str.lower().str.contains('lpt')].index, inplace=True)
ulpt_df.drop(ulpt_df[ulpt_df['selftext'].str.lower().str.contains('lpt')].index, inplace=True)

lpt_df.shape

ulpt_df.shape

Let's verify if there is actually any lpt anywhere in the title or text:

lpt_df[lpt_df['title'].str.contains(r'(?i)lpt')]

lpt_df[lpt_df['selftext'].str.contains(r'(?i)lpt')]

ulpt_df[ulpt_df['title'].str.contains(r'(?i)lpt')]

#Dropping all the 'u's with spaces in case there was a ulpt somewhere in there that got cut to just u
ulpt_df.drop(ulpt_df[ulpt_df['selftext'].str.contains(r'(?i) u ')].index, inplace=True)

len(ulpt_df)

Let's inspect our new dataframes:

lpt_df.head()

ulpt_df.head()

lpt_df.to_csv('./data/lpt_cleaned.csv')
ulpt_df.to_csv('./data/ulpt_cleaned.csv')
both_df = pd.concat([lpt_df, ulpt_df])
both_df.reset_index(inplace=True)
both_df.head()

both_df.drop('index', axis=1, inplace=True)

both_df.info()

both_df.to_csv('./data/both_cleaned.csv')

Great. Now we have separate clean dataframes ready for analysis and modeling. Please Continue on to notebook 03.
